refactor(components): rename Preview/SourceView files to kebab-case
Aligns Preview/SourceView with the kebab-case convention introduced by
the file-tree split in #15. Updates the two App.tsx imports and the
dynamic `await import("./SourceView")` / static `./Preview` import
inside the test files.

Closes #16
