+++ frontend/src/App.tsx
@@ -3,8 +3,8 @@ import { fetchFiles, fetchFileSource, type FileEntry } from "./api/files";
 import { subscribe } from "./api/events";
 import { renderPlantUML } from "./plantuml/renderer";
 import { FileTree } from "./components/file-tree";
-import { Preview } from "./components/Preview";
-import { SourceView } from "./components/SourceView";
+import { Preview } from "./components/preview";
+import { SourceView } from "./components/source-view";
 import { SOURCE_PANEL_ID, SOURCE_PANEL_NAME, SOURCE_TOGGLE_LABEL } from "./sourcePanel";
 
 type RenderState =
