+++ frontend/src/components/preview.tsx
